Struct ebay_easy_api::search::Browser

pub struct Browser<'c>(/* private fields */);

A Browser API client.

This structure references an instance of EbayApiClient, so it can only be used as long as the referenced EbayApiClient lives.

Implementations

impl<'c> Browser<'c>

pub fn search<S: AsRef<str>>( &self, query: S, limit: usize, ) -> Result<SearchResults, Error>

Perform a search using the given query string and return up to limit results.

Errors

This can return an error if the request fails, or the response body could not be deserialized.

Example

let client = EbayApiClient::new_unchecked(token, marketplace);

// Obtain a search API client
let browser = client.search();

// Define a limit
let limit: usize = 3;

// Perform the search query
let results = browser.search("gaming pc", limit);

// If the search was successfull, we should get the results.
assert!(results.is_ok());

// We should not get more than `limit` results.
assert!(results.unwrap().len() <= limit);

pub fn find_item<S: AsRef<str>>(&self, id: S) -> Result<Option<Item>, Error>

Finds an item by it’s ID.

Errors

This can return an error if the request fails or the response body could not be deserialized. Furthermore, if the item was not found None is returned.

Example

let client = EbayApiClient::new_unchecked(token, marketplace);

// Obtain a search API client
let browser = client.search();

// Perform the search query
let result = browser.find_item("v1|202975928242|0");

// If the search was successfull, we should get the results.
assert!(result.is_ok());

match result.unwrap() {
    Some(item) => println!("Found {}, with price {}", item.title, item.price),
    None => println!("Item was not found"),
}